+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Add AuthMe database schema support (Fixes #49)
 * Add /flexiblelogin forcelogin subcommand.
+* Add /l to /login command alias. (can be disabled with supportSomeChatPlugins=true)
+* Add /fl to /flexiblelogin command alias.
+* Add /cp to /changepassword command alias.
 
 ###### 0.17